fix(linewrap): Linewrap doesn't require an additional argument
Given `w` is simply a flag not a option in the sense that it requires an
additional argument. `getopts` opts suffixed with a `:` require an additional arg.[1]

```quote
If a character is
followed by a <colon>, the option shall be expected to
have an argument, which should be supplied as a
separate argument.
```

[1]: https://man7.org/linux/man-pages/man1/getopts.1p.html